like the history of a workflow job
